requirements.txt
zeep
spyne
flask
requests
graphene
flask_graphql
graphql-core==2.3.1
grpcio
grpcio-tools